examples: add a DepthSlice and Depth materials for debugging
